requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.5
aiosignal==1.3.2
astroid==3.3.8
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.6.10
dacite==1.9.2
dill==0.3.9
docker==7.1.0
frozenlist==1.5.0
idna==3.10
isort==6.0.0
mccabe==0.7.0
multidict==6.1.0
platformdirs==4.3.6
propcache==0.2.1
pylint==3.3.4
python-dotenv==1.0.1
requests==2.32.3
tomlkit==0.13.2
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
